+++ DocumentationTests/DocumentationTests/DocumentationTests/05-ui-cookbook/05-incoming-call.swift
@@ -80,20 +80,11 @@ fileprivate func content() {
 
                 }
                 .background(Color.white.edgesIgnoringSafeArea(.all))
-                .onChange(of: viewModel.hideIncomingCallScreen) { newValue in
-                    if newValue {
-                        callViewModel.rejectCall(callType: callInfo.type, callId: callInfo.id)
-                    }
-                }
-                .onDisappear {
-                    viewModel.stopTimer()
-                }
             }
 
             var callInfo: IncomingCall {
                 viewModel.callInfo
             }
-
         }
 
         class CustomViewFactory: ViewFactory {